+++ dart/lib/src/http_client/failed_request_client.dart
@@ -126,8 +126,7 @@ class FailedRequestClient extends BaseClient {
       } else if (failedRequestStatusCodes.containsStatusCode(statusCode)) {
         // Capture an exception if the status code is considered bad
         capture = true;
-        reason =
-            'Event was captured because the request status code was $statusCode';
+        reason = 'HTTP Client Error with status code: $statusCode';
         exception ??= SentryHttpClientError(reason);
       }
       if (capture) {
@@ -184,11 +183,22 @@ class FailedRequestClient extends BaseClient {
       type: 'SentryHttpClient',
       description: reason,
     );
-    final throwableMechanism = ThrowableMechanism(mechanism, exception);
+
+    bool? snapshot;
+    if (exception is SentryHttpClientError) {
+      snapshot = true;
+    }
+
+    final throwableMechanism = ThrowableMechanism(
+      mechanism,
+      exception,
+      snapshot: snapshot,
+    );
 
     final event = SentryEvent(
       throwable: throwableMechanism,
       request: sentryRequest,
+      timestamp: _hub.options.clock(),
     );
 
     if (response != null) {