+++ dart/lib/src/sentry_tracer.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:intl/intl.dart';
 import 'package:meta/meta.dart';
-import 'utils.dart';
 
 import '../sentry.dart';
 import 'sentry_tracer_finish_status.dart';
@@ -72,7 +71,7 @@ class SentryTracer extends ISentrySpan {
 
   @override
   Future<void> finish({SpanStatus? status, DateTime? endTimestamp}) async {
-    final commonEndTimestamp = endTimestamp ?? getUtcDateTime();
+    final commonEndTimestamp = endTimestamp ?? _hub.options.clock();
     _autoFinishAfterTimer?.cancel();
     _finishStatus = SentryTracerFinishStatus.finishing(status);
     if (!_rootSpan.finished &&