+++ dart/lib/src/throwable_mechanism.dart
@@ -4,10 +4,17 @@ import 'protocol/mechanism.dart';
 class ThrowableMechanism implements Exception {
   final Mechanism _mechanism;
   final dynamic _throwable;
+  final bool? _snapshot;
 
-  ThrowableMechanism(this._mechanism, this._throwable);
+  ThrowableMechanism(
+    this._mechanism,
+    this._throwable, {
+    bool? snapshot,
+  }) : _snapshot = snapshot;
 
   Mechanism get mechanism => _mechanism;
 
   dynamic get throwable => _throwable;
+
+  bool? get snapshot => _snapshot;
 }