+++ flutter/lib/src/renderer/renderer.dart
@@ -1,19 +1,26 @@
+import 'package:meta/meta.dart';
+
 import 'unknown_renderer.dart'
     if (dart.library.html) 'html_renderer.dart'
     if (dart.library.io) 'io_renderer.dart' as implementation;
 
-FlutterRenderer getRenderer() => implementation.getRenderer();
+@internal
+class RendererWrapper {
+  FlutterRenderer getRenderer() {
+    return implementation.getRenderer();
+  }
 
-String getRendererAsString() {
-  switch (getRenderer()) {
-    case FlutterRenderer.skia:
-      return 'Skia';
-    case FlutterRenderer.canvasKit:
-      return 'CanvasKit';
-    case FlutterRenderer.html:
-      return 'HTML';
-    case FlutterRenderer.unknown:
-      return 'Unknown';
+  String getRendererAsString() {
+    switch (getRenderer()) {
+      case FlutterRenderer.skia:
+        return 'Skia';
+      case FlutterRenderer.canvasKit:
+        return 'CanvasKit';
+      case FlutterRenderer.html:
+        return 'HTML';
+      case FlutterRenderer.unknown:
+        return 'Unknown';
+    }
   }
 }